move lock files to npm manager
