+++ lib/manager/npm/post-update/index.js
@@ -10,7 +10,7 @@ module.exports = {
   determineLockFileDirs,
   writeExistingFiles,
   writeUpdatedPackageFiles,
-  getUpdatedLockFiles,
+  getAdditionalFiles,
 };
 
 // Strips empty values, deduplicates, and returns the directories from filenames
@@ -291,8 +291,8 @@ async function writeUpdatedPackageFiles(config) {
 }
 
 // istanbul ignore next
-async function getUpdatedLockFiles(config, packageFiles) {
-  logger.trace({ config }, 'getUpdatedLockFiles');
+async function getAdditionalFiles(config, packageFiles) {
+  logger.trace({ config }, 'getAdditionalFiles');
   logger.debug('Getting updated lock files');
   const lockFileErrors = [];
   const updatedLockFiles = [];
