+++ lib/manager/npm/post-update/npm.js
