+++ lib/manager/npm/post-update/pnpm.js
