+++ lib/workers/branch/index.js
@@ -1,6 +1,6 @@
 const schedule = require('./schedule');
 const { getUpdatedPackageFiles } = require('./get-updated');
-const { getUpdatedLockFiles } = require('./lock-files');
+const { getAdditionalFiles } = require('../../manager/npm/post-update');
 const { commitFilesToBranch } = require('./commit');
 const { getParentBranch } = require('./parent');
 const { tryBranchAutomerge } = require('./automerge');
@@ -143,7 +143,7 @@ async function processBranch(branchConfig, packageFiles) {
     } else {
       logger.debug('No package files need updating');
     }
-    Object.assign(config, await getUpdatedLockFiles(config, packageFiles));
+    Object.assign(config, await getAdditionalFiles(config, packageFiles));
     if (config.updatedLockFiles && config.updatedLockFiles.length) {
       logger.debug(
         { updatedLockFiles: config.updatedLockFiles.map(f => f.name) },
